* change int values in POJO to long * add catch exception in send and map OrderCreateRequest * create lower range of change price and qty

+++ src/main/java/me/ikosarim/cripto_bot/db_model/TradeHistoryEntity.java
@@ -17,7 +17,7 @@ public class TradeHistoryEntity {
 
     @Id
     @Column(name = "trade_id", nullable = false)
-    private Integer tradeId;
+    private Long tradeId;
 
     @Column(name = "date", nullable = false)
     private Date date;
